fix architecture tests
