+++ src/test/java/de/tum/cit/aet/artemis/core/service/CourseAdminServiceTest.java
@@ -25,7 +25,7 @@ import de.tum.cit.aet.artemis.programming.repository.ProgrammingExerciseReposito
 import de.tum.cit.aet.artemis.programming.util.ProgrammingExerciseFactory;
 import de.tum.cit.aet.artemis.shared.base.AbstractSpringIntegrationLocalCILocalVCTest;
 
-public class CourseAdminServiceTest extends AbstractSpringIntegrationLocalCILocalVCTest {
+class CourseAdminServiceTest extends AbstractSpringIntegrationLocalCILocalVCTest {
 
     private static final String TEST_PREFIX = "courseadminservice";
